refactor: split connect defines
