+++ packages/connection/src/browser/connect.ts
@@ -1,0 +1,9 @@
+import {
+  BrowserMessageReader,
+  BrowserMessageWriter,
+  createMessageConnection,
+} from '@opensumi/vscode-jsonrpc/lib/browser/main';
+
+export function createSocketConnection(socket: Worker) {
+  return createMessageConnection(new BrowserMessageReader(socket), new BrowserMessageWriter(socket));
+}
