+++ packages/connection/src/browser/index.ts
@@ -1,0 +1,2 @@
+export * from './ws-channel-handler';
+export * from './connect';
