+++ packages/connection/src/common/index.ts
@@ -3,3 +3,4 @@ export * from './proxy';
 export * from './rpcProtocol';
 export * from './utils';
 export * from './ws-channel';
+export * from './connect';
