+++ packages/connection/src/node/connect.ts
@@ -1,0 +1,11 @@
+import {
+  SocketMessageReader,
+  SocketMessageWriter,
+  createMessageConnection,
+} from '@opensumi/vscode-jsonrpc/lib/node/main';
+
+import net from 'net';
+
+export function createSocketConnection(socket: net.Socket) {
+  return createMessageConnection(new SocketMessageReader(socket), new SocketMessageWriter(socket));
+}
