+++ packages/connection/src/node/index.ts
@@ -1,3 +1,7 @@
+import { SocketMessageReader, SocketMessageWriter } from '@opensumi/vscode-jsonrpc/lib/node/main';
+
 export * from './ws';
 export * from './common-channel-handler';
-export * from './center';
+export * from './connect';
+
+export { SocketMessageReader, SocketMessageWriter };
