+++ packages/core-browser/src/bootstrap/connection.ts
@@ -1,10 +1,5 @@
-import {
-  RPCServiceCenter,
-  initRPCService,
-  WSChannelHandler,
-  createSocketConnection,
-  RPCMessageConnection,
-} from '@opensumi/ide-connection';
+import { RPCServiceCenter, initRPCService, RPCMessageConnection } from '@opensumi/ide-connection';
+import { WSChannelHandler, createSocketConnection } from '@opensumi/ide-connection/lib/browser';
 import { createWebSocketConnection } from '@opensumi/ide-connection/lib/common/message';
 import { Injector, Provider } from '@opensumi/di';
 import {
