+++ packages/debug/__tests__/browser/view/console/debug-console.view.test.tsx
@@ -30,7 +30,7 @@ import {
 import { IWorkspaceService } from '@opensumi/ide-workspace';
 import { QuickPickService } from '@opensumi/ide-quick-open';
 import { IEditorDocumentModelService } from '@opensumi/ide-editor/lib/browser';
-import { WSChannelHandler } from '@opensumi/ide-connection';
+import { WSChannelHandler } from '@opensumi/ide-connection/lib/browser/ws-channel-handler';
 import { IVariableResolverService } from '@opensumi/ide-variable';
 import { ITaskService } from '@opensumi/ide-task';
 import { DebugConsoleFilterService } from '@opensumi/ide-debug/lib/browser/view/console/debug-console-filter.service';
