+++ packages/extension/__tests__/browser/extension-service/extension.contribution.test.ts
@@ -2,8 +2,8 @@ import { MockInjector, mockService } from '@opensumi/ide-dev-tool/src/mock-injec
 import { IMainLayoutService } from '@opensumi/ide-main-layout';
 
 import { setupExtensionServiceInjector } from './extension-service-mock-helper';
-import { getClientId } from '../../../lib/browser/extension.contribution';
-import { WSChannelHandler } from '@opensumi/ide-connection';
+import { getClientId } from '../../../src/browser/extension.contribution';
+import { WSChannelHandler } from '@opensumi/ide-connection/lib/browser/ws-channel-handler';
 import { uuid } from '@opensumi/ide-core-common';
 
 describe('Extension service', () => {
