+++ packages/extension/src/node/extension.host.proxy.manager.ts
@@ -11,7 +11,8 @@ import {
 } from '../common';
 import net from 'net';
 import { RPCServiceCenter, INodeLogger } from '@opensumi/ide-core-node';
-import { createSocketConnection, getRPCService, RPCProtocol, IRPCProtocol } from '@opensumi/ide-connection';
+import { getRPCService, RPCProtocol, IRPCProtocol } from '@opensumi/ide-connection';
+import { createSocketConnection } from '@opensumi/ide-connection/lib/node';
 
 @Injectable()
 export class ExtensionHostProxyManager implements IExtensionHostManager {
