+++ packages/remote-opener/src/browser/remote.opener.contribution.ts
@@ -9,7 +9,7 @@ import {
 } from '../common';
 
 import { AppConfig, ClientAppContribution, electronEnv } from '@opensumi/ide-core-browser';
-import { WSChannelHandler } from '@opensumi/ide-connection/lib/browser/ws-channel-handler';
+import { WSChannelHandler } from '@opensumi/ide-connection/lib/browser';
 // 从extension.contribution.ts中Copy过来，因为直接引入会有一定概率触发IDE初始化问题
 const getClientId = (injector: Injector) => {
   let clientId: string;
