+++ tools/electron/src/browser/app.tsx
@@ -1,9 +1,8 @@
 /* eslint-disable no-console */
 console.time('Render');
-import type { Socket } from 'net';
 import { ClientApp, IClientAppOpts, electronEnv, URI } from '@opensumi/ide-core-browser';
 import { Injector, Domain } from '@opensumi/di';
-import { createSocketConnection } from '@opensumi/ide-connection';
+import { createSocketConnection } from '@opensumi/ide-connection/lib/browser';
 
 // 引入公共样式文件
 import '@opensumi/ide-core-browser/lib/style/index.less';
@@ -68,5 +67,5 @@ export async function renderApp(arg1: IClientAppOpts | Domain, arg2: Domain[] = 
 }
 
 declare global {
-  function createRPCNetConnection(): Promise<Socket>;
+  function createRPCNetConnection(): Promise<Worker>;
 }
